Add a 700+ starred stupid trash.
https://github.com/Ha3MrX/DDos-Attack
You should try this shitty script I just found.
"It's fake, I think it's a joke. Look at the code. Well everyone using this is prob like 5 years old and don't even know how to write a head title in HTML lol"

"Thanks Ha3MrX,
Ive been looking for DDOS tools for so so so long and all the shit i try and get are a virus or fake but thanks this script works and the source code is so easy to install ;)
I cant wait to use it some more, if people talk shit well.... Thanks to u i can hit people with 169Gb/per second. but lol i love the 69 but im so glad this powerful took works."